pyproject.toml
[tool.poetry]
name = "lab"
version = "0.1.0"
description = ""
authors = ["Antonis Tziorvas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
pandas = "^2.2.1"
geopandas = "^0.14.3"
matplotlib = "^3.8.4"
seaborn = "^0.13.2"
scikit-learn = "^1.4.1.post1"
ipykernel = "^6.29.4"
requests = "^2.31.0"
folium = "^0.16.0"
mapclassify = "^2.6.1"
shapely = "^2.0.3"
xgboost = "^2.0.3"
nbformat = "^5.10.4"
lightgbm = "^4.4.0"
bokeh = "^3.4.2"
scikit-optimize = "^0.10.2"
tqdm = "^4.66.4"
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
sqlalchemy-utils = "^0.41.2"
pyarrow = "^17.0.0"
fastparquet = "^2024.5.0"
streamlit-folium = "^0.22.0"
sshtunnel = "^0.4.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
mypy = "^1.9.0"
isort = "^5.13.2"
tqdm = "^4.66.4"
ipywidgets = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
